some cleanups in preparation for adding new functionalities
